requirements.txt
requests==2.31.0
python-dotenv==1.0.0
loguru==0.6.0
fastapi==0.95.2
pydantic==1.10.5
mysql-connector-python==8.0.33
schedule==1.1.0
uvicorn==0.23.0
pika==1.3.2
